--- Website/Anotai/Views/Noticias/Delete.cshtml
@@ -1,48 +1,0 @@
-﻿@model Anotai.Models.Noticia
-
-@{
-    Layout = null;
-}
-
-<!DOCTYPE html>
-
-<html>
-<head>
-    <meta name="viewport" content="width=device-width" />
-    <title>Delete</title>
-</head>
-<body>
-    <h3>Are you sure you want to delete this?</h3>
-    <div>
-        <h4>Noticia</h4>
-        <hr />
-        <dl class="dl-horizontal">
-            <dt>
-                @Html.DisplayNameFor(model => model.Titulo)
-            </dt>
-    
-            <dd>
-                @Html.DisplayFor(model => model.Titulo)
-            </dd>
-    
-            <dt>
-                @Html.DisplayNameFor(model => model.Mensagem)
-            </dt>
-    
-            <dd>
-                @Html.DisplayFor(model => model.Mensagem)
-            </dd>
-    
-        </dl>
-    
-        @using (Html.BeginForm()) {
-            @Html.AntiForgeryToken()
-    
-            <div class="form-actions no-color">
-                <input type="submit" value="Delete" class="btn btn-default" /> |
-                @Html.ActionLink("Back to List", "Index")
-            </div>
-        }
-    </div>
-</body>
-</html>
