--- Website/Anotai/Views/Noticias/Details.cshtml
@@ -1,42 +1,0 @@
-﻿@model Anotai.Models.Noticia
-
-@{
-    Layout = null;
-}
-
-<!DOCTYPE html>
-
-<html>
-<head>
-    <meta name="viewport" content="width=device-width" />
-    <title>Details</title>
-</head>
-<body>
-    <div>
-        <h4>Noticia</h4>
-        <hr />
-        <dl class="dl-horizontal">
-            <dt>
-                @Html.DisplayNameFor(model => model.Titulo)
-            </dt>
-    
-            <dd>
-                @Html.DisplayFor(model => model.Titulo)
-            </dd>
-    
-            <dt>
-                @Html.DisplayNameFor(model => model.Mensagem)
-            </dt>
-    
-            <dd>
-                @Html.DisplayFor(model => model.Mensagem)
-            </dd>
-    
-        </dl>
-    </div>
-    <p>
-        @Html.ActionLink("Edit", "Edit", new { id = Model.NoticiaId }) |
-        @Html.ActionLink("Back to List", "Index")
-    </p>
-</body>
-</html>
