--- Website/Anotai/Views/Noticias/Index.cshtml
@@ -1,47 +1,0 @@
-﻿@model IEnumerable<Anotai.Models.Noticia>
-
-@{
-    Layout = null;
-}
-
-<!DOCTYPE html>
-
-<html>
-<head>
-    <meta name="viewport" content="width=device-width" />
-    <title>Index</title>
-</head>
-<body>
-    <p>
-        @Html.ActionLink("Create New", "Create")
-    </p>
-    <table class="table">
-        <tr>
-            <th>
-                @Html.DisplayNameFor(model => model.Titulo)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.Mensagem)
-            </th>
-            <th></th>
-        </tr>
-    
-    @foreach (var item in Model) {
-        <tr>
-            <td>
-                @Html.DisplayFor(modelItem => item.Titulo)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.Mensagem)
-            </td>
-            <td>
-                @Html.ActionLink("Edit", "Edit", new { id=item.NoticiaId }) |
-                @Html.ActionLink("Details", "Details", new { id=item.NoticiaId }) |
-                @Html.ActionLink("Delete", "Delete", new { id=item.NoticiaId })
-            </td>
-        </tr>
-    }
-    
-    </table>
-</body>
-</html>
